% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vonMises.R
\name{besselexp}
\alias{besselexp}
\alias{dbesselexp}
\alias{dbesselexpkern}
\alias{rbesselexp}
\title{Bessel exponential distribution}
\usage{
dbesselexp(kp, eta = 1, g = -0.5, log = FALSE)

dbesselexpkern(kp, eta, g, log = FALSE)

rbesselexp(n, eta, g)
}
\arguments{
\item{kp}{Numeric;}

\item{eta}{Integer; This is the posterior sample size, which is n + c where c
is the number of observations contained in the conjugate prior. For
uninformative, \code{c = 0} and \code{eta = n}.}

\item{g}{Numeric; Should be \code{-R*cos(mu-theta_bar)/eta}, where \code{R}
is the posterior mean resultant length, and \code{theta_bar} is the
posterior mean, while \code{mu} is the current value of the mean. Note that
this parameter is called \eqn{\beta_0} in Forbes and Mardia (2015).}
}
\value{
For \code{dbesselexp} and \code{dbesselexpkern}, a scalar. For
  \code{rbesselexp}, a vector of random variates from the distribution.
}
\description{
Functions for the Bessel exponential distribution. This is the conditional
posterior distribution of the concentration parameter \code{kappa} of a von
Mises distribution with conjugate prior. The random generation algorithm is
due to Forbes and Mardia (2015).
}
\section{Functions}{
\itemize{
\item \code{dbesselexp}: Probability density function.

\item \code{dbesselexpkern}: Kernel (unnormalized version) of the pdf.

\item \code{rbesselexp}: Random generation.
}}

\examples{
dbesselexp(2, 20, -.5)
plot(density(rbesselexp(100, 20, -.5)), xlim = c(0, 5))

# Plot probability density function
dbesexpfun <- Vectorize(function(x) dbesselexp(x, 20, -.5))
curve(dbesexpfun, 0, 5)

# Compare with density of random draws
plot(density(rbesselexp(1000, 20, -.5)), xlim = c(0, 5))

}
