% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{BIC.batmixmod}
\alias{BIC.batmixmod}
\title{Extract BIC from Batmixmod}
\usage{
\method{BIC}{batmixmod}(bm_mod, ...)
}
\arguments{
\item{bm_mod}{A \code{batmixmod} object.}

\item{...}{Ignored.}
}
\value{
The numerical value of the BIC.
}
\description{
Extract BIC from Batmixmod
}
