% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapEMBatMix.R
\name{bootstrapEMBatMix}
\alias{bootstrapEMBatMix}
\title{Bootstrap the Batschelet mixture parameters}
\usage{
bootstrapEMBatMix(x, B = 500, parallel = TRUE, verbose = FALSE,
  bat_type = "power", n_comp = 4, init_pmat = matrix(NA, n_comp, 4),
  fixed_pmat = matrix(NA, n_comp, 4), ll_tol = 0.1, max_its = 50,
  optimization_its = 10)
}
\arguments{
\item{x}{Numeric; A set angles in radians.}

\item{B}{Integer; The number of bootstrap samples.}

\item{parallel}{Logical; Whether to perform bootstraps in parallel.}

\item{verbose}{Logical; Whether to print debug info.}

\item{bat_type}{Either 'inverse' or 'power', the type of distribution to fit.
The two distributions are similar, but the power Batschelet distribution is
computationally much less demanding.}

\item{n_comp}{Integer; Fixed number of components to estimate.}

\item{init_pmat}{A numeric matrix with \code{n_comp} rows and four columns,
corresponding to \eqn{\mu, \kappa, \lambda, \alpha}, in that order. Gives
starting values for the parameters. If any element is \code{NA}, it will be
given a default starting value. For \eqn{mu}, the default starting values
are equally spaced on the circle. For \eqn{\kappa}, the default starting
value is 5. For \eqn{\lambda}, the default starting value is 0, which
corresponds to the von Mises distribution. For \eqn{\alpha}, the default
starting value is \code{1/n_comp}.}

\item{fixed_pmat}{A numeric matrix with \code{n_comp} rows and four columns,
corresponding to \eqn{\mu, \kappa, \lambda, \alpha}, in that order. Any
element that is not \code{NA} in this matrix will be held constant at the
given value and not sampled.}

\item{ll_tol}{The likelihood tolerance for the EM-algorithm.}

\item{max_its}{The maximum number of E-M iterations for the EM-algorithm.}

\item{optimization_its}{The maximum number of maximization iterations within the EM-algorithm.}
}
\value{
A matrix of sampled bootstrap values.
}
\description{
Bootstrap the Batschelet mixture parameters
}
\examples{
x <- rinvbatmix(100)

 bootstrapEMBatMix(x, B = 5, parallel = FALSE)

}
