% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{bridge_sampler.batmixmod}
\alias{bridge_sampler.batmixmod}
\title{Log Marginal Likelihood via Bridge Sampling}
\usage{
\method{bridge_sampler}{batmixmod}(bm_mod, ...)
}
\arguments{
\item{bm_mod}{A \code{batmixmod} object.}

\item{...}{Additional arguments to be passed to \code{bridge_sampler.matrix}}
}
\description{
More details are provided under
  \code{\link[bridgesampling:bridge_sampler]{bridge_sampler}}.
}
\seealso{
\code{
  \link[brms:bayes_factor]{bayes_factor},
  \link[brms:post_prob]{post_prob}
}
}
