% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movMF.R
\name{coef_polar_movMF}
\alias{coef_polar_movMF}
\title{Compute the polar versions of the movMF results}
\usage{
coef_polar_movMF(m)
}
\arguments{
\item{m}{A movMF object.}
}
\value{
A matrix of mus and kps.
}
\description{
The results of a movMF object are given as a vector called theta. We move to polar coordinates with
mu and kp.
}
