% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsOfBatParameters.R
\name{computeMeanResultantLengthBat}
\alias{computeMeanResultantLengthBat}
\title{The mean resultant length of a Batschelet-type distribution.}
\usage{
computeMeanResultantLengthBat(kp, lam, dbat_fun = dinvbat)
}
\arguments{
\item{kp}{The concentration parameter.}

\item{lam}{The peakedness parameter.}

\item{dbat_fun}{The pdf function of the required Batschelet distribution.}
}
\value{
The mean resultant length.
}
\description{
The mean resultant length of a Batschelet-type distribution.
}
