% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsOfBatParameters.R
\name{computeMeanResultantLengthVM}
\alias{computeMeanResultantLengthVM}
\title{The mean resultant length of a von Mises distribution.}
\usage{
computeMeanResultantLengthVM(kp)
}
\arguments{
\item{kp}{The concentration parameter.}
}
\value{
The mean resultant length.
}
\description{
The mean resultant length of a von Mises distribution.
}
