% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InvBatschelet.R
\name{dinvbat}
\alias{dinvbat}
\alias{rinvbat}
\title{The von-Mises based symmetric inverse Batschelet distribution}
\usage{
dinvbat(x, mu = 0, kp = 1, lam = 0, log = FALSE)

rinvbat(n, mu = 0, kp = 1, lam = 0)
}
\arguments{
\item{x}{An angle in radians.}

\item{mu}{A mean direction, in radians.}

\item{kp}{Numeric, \eqn{> 0,}the concentration parameter.}

\item{lam}{The shape parameter (peakedness), -1 < \code{lam} < 1.}

\item{log}{Logical; whether to return the log of the probability or not.}

\item{n}{The number of random variates to obtain.}
}
\value{
Numeric. For \code{dinvbat}, the probability or log-probability of
  angle x given the parameters. For \code{rinvbat}, a vector of random
  variates from the inverse Batschelet distribution.
}
\description{
The inverse Batschelet distribution, with mean direction \code{mu},
concentration parameter \code{kp}, and shape (peakedness) parameter
\code{lam}. This is the von Mises based version, without a skewness
parameter.
}
\section{Functions}{
\itemize{
\item \code{rinvbat}: Random generation
}}

\examples{
dinvbat(3)

# Peaked distribution
curve(dinvbat(x, lam = .8), -pi, pi)

# Flat-topped distribution
curve(dinvbat(x, lam = -.8), -pi, pi)

}
