% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PowBatschelet.R
\name{dpowbat}
\alias{dpowbat}
\title{The von-Mises based symmetric power Batschelet distribution}
\usage{
dpowbat(x, mu = 0, kp = 1, lam = 0, log = FALSE)
}
\arguments{
\item{x}{An angle in radians.}

\item{mu}{A mean direction, in radians.}

\item{kp}{Numeric, \eqn{> 0,}the concentration parameter.}

\item{lam}{The shape parameter (peakedness), -Inf < \code{lam} < Inf}

\item{log}{Logical; whether to return the log of the probability or not.}
}
\value{
Numeric; Either the probability or log-probability of angle x given the parameters.
}
\description{
The power Batschelet distribution, with mean direction \code{mu}, concentration parameter \code{kp}, and
shape (peakedness) parameter \code{lam}. This is the von Mises based version, without a skewness
parameter.
}
\examples{
dpowbat(3)

# Flat-topped distribution
curve(dpowbat(x, lam = -.8), -pi, pi)

# Peaked distribution
curve(dpowbat(x, lam = .8), -pi, pi)

}
