% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PowBatschelet.R
\name{dpowbatkern}
\alias{dpowbatkern}
\title{Kernel of the von-Mises based symmetric power Batschelet distribution}
\usage{
dpowbatkern(x, mu = 0, kp = 1, lam = 0, log = FALSE)
}
\value{
The unnormalized density value of the power Batschelet distribution.
}
\description{
Kernel of the von-Mises based symmetric power Batschelet distribution
}
