% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vonMises.R
\name{dvm}
\alias{dvm}
\title{Compute the density of the von Mises distribution}
\usage{
dvm(x, mu = 0, kp = 1, log = FALSE)
}
\arguments{
\item{x}{An angle in radians for which to compute the pdf.}

\item{mu}{The mean direction.}

\item{kp}{The concentration parameter, \eqn{kappa}.}

\item{log}{Logical; If TRUE, return the log of the probability of x.}
}
\description{
Compute the probability density function of the von Mises distribution.
}
\examples{
dvm(3)
dvm(3, -2.2, 10)
}
