% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatMixturesFitting.R
\name{fitbatmix}
\alias{fitbatmix}
\title{Fit a mixture of Batschelet distributions}
\usage{
fitbatmix(x, method = "bayes", bat_type = "power", n_comp = 4,
  init_pmat = matrix(NA, n_comp, 4), fixed_pmat = matrix(NA, n_comp, 4),
  probs = c(0.025, 0.975), post_est_median = TRUE, ...)
}
\arguments{
\item{x}{A dataset of angles in radians.}

\item{method}{Character; One of "bayes", "EM", or "boot". The method of
obtaining a fit.}

\item{bat_type}{Character; Either "power" or "inverse", denoting the type of
Batschelet distribution to employ.}

\item{n_comp}{The number of components to be used in the mixture. This is
fixed, so it can not be estimated.}

\item{init_pmat}{An \code{n_comp * 4} matrix, the initial values of the
parameter matrix. The parameters are ordered \code{mu}, \code{kp},
\code{lam} and then \code{alph}.}

\item{fixed_pmat}{An \code{n_comp * 4} matrix, containing a parameter matrix,
with \code{NA} for parameters to be estimated, and a numeric for each
parameter that should be kept fixed to a specific value.}

\item{probs}{Numeric vector; The probabilities for which to compute quantiles
in summarizing bootstrap or MCMC samples. By default, \code{probs = c(.025,
.975)}, which corresponds to standard 95\% confidence or credible
intervals.}

\item{post_est_median}{Logical; Only relevant for MCMC. Whether to use the
posterior median as the estimate. If FALSE (the default) we use the mean,
or mean direction for \code{mu}.}

\item{...}{Additional arguments to be passed to the selected \code{method}.
In particular, use \code{verbose = TRUE} to print debug statements.}
}
\value{
An object of class 'batmixmod'.
}
\description{
This is the main function of the package \code{flexcircmix}, and functions as
an interface to fit mixtures of Batschelet-type distributions, using
frequentist or Bayesian methods.
}
\examples{
x <- rinvbatmix(50)
fitbatmix(x, method = "EM")

}
