% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InvBatschelet.R
\name{force_neg_pi_pi}
\alias{force_neg_pi_pi}
\title{Set an angle to have its bounds in (-pi, pi)}
\usage{
force_neg_pi_pi(x)
}
\arguments{
\item{x}{Numeric, representing an angle.}
}
\value{
Numeric, representing an angle, between -pi and pi.
}
\description{
This function is used to make sure that an angle is represented as being between -pi and pi.
}
