% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movMF.R
\name{invbatmix_pmat_from_movMF}
\alias{invbatmix_pmat_from_movMF}
\title{Obtain the inverse Batschelet parameter matrix from a movMF object}
\usage{
invbatmix_pmat_from_movMF(m)
}
\arguments{
\item{m}{A movMF object.}
}
\value{
A parameter matrix with inverse Batschelet mixture parameters.
}
\description{
This function allows using the more general inverse Batschelet functions, such as plotting. The
results of a movMF object are given as a vector called theta. We move to polar coordinates with
mu and kp, set lambda to 0 (the von Mises distribution as a special case of the inverse
Batschelet), and obtain the alphas from the movMF object.
}
