% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InvBatschelet.R
\name{K_kplam}
\alias{K_kplam}
\title{Compute a partial normalizing constant of a vM-based inverse Batschelet distribution}
\usage{
K_kplam(kp, lam)
}
\arguments{
\item{kp}{Numeric, \eqn{> 0,}the concentration parameter.}

\item{lam}{The shape parameter (peakedness), -1 < \code{lam} < 1.}
}
\value{
Numeric.
}
\description{
The value given is only the additional part of the inverse Batschelet. The von Mises normalizing
constant must still be added.
}
