% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcBatschelet.R
\name{lam_beta_log_prior_2_2}
\alias{lam_beta_log_prior_2_2}
\title{A rescaled beta prior for lambda}
\usage{
lam_beta_log_prior_2_2(lam)
}
\arguments{
\item{lam}{Numeric;}
}
\value{
The log-prior probability.
}
\description{
This prior is symmetric between -1 and 1, and captures the prior belief that
values of \code{lam} on the boundary of the parameter space are a prior
unlikely.
}
