% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InvBatschelet.R
\name{likfuninvbat}
\alias{likfuninvbat}
\title{@describeIn likfuninvbat Return a likelihood function.}
\usage{
likfuninvbat(x, weights = rep(1, length(x)), log = TRUE)
}
\description{
@describeIn likfuninvbat Return a likelihood function.
}
