% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PowBatschelet.R
\name{likpowbat}
\alias{likpowbat}
\alias{likfunpowbat}
\title{Likelihood of the power Batschelet distribution}
\usage{
likpowbat(x, mu, kp, lam, weights = rep(1, length(x)), log = TRUE)

likfunpowbat(x, weights = rep(1, length(x)), log = TRUE)
}
\arguments{
\item{x}{A set of angles in radians.}

\item{mu}{A mean direction, in radians.}

\item{kp}{Numeric, \eqn{> 0,}the concentration parameter.}

\item{lam}{The shape parameter (peakedness), -1 < \code{lam} < 1.}

\item{weights}{A vector of length \code{length(x)}, which gives importace
weights to be used for \code{x}.}

\item{log}{If \code{TRUE} (the default), the log-likelihood is used.}
}
\value{
\code{likpowbat} returns a value, the likelihood given the data and
  parameters. \code{likfunpowbat} returns a function of mu, kp and lam, which
  can be evaluated later for a given set of parameters.
}
\description{
Two functions to obtain the power Batschelet likelihood of a set parameters,
given a data set of angles.
}
\section{Functions}{
\itemize{
\item \code{likfunpowbat}: Return a likelihood function.
}}

\examples{
vm_data <- circglmbayes::rvmc(10, 1, 5)
llfun <- likfunpowbat(vm_data)

# log-likelihood value of true parameters.
llfun(mu = 1, kp = 5, lam = 0)

# Plot the conditional log-likelihood.
kp_conditional_ll <- Vectorize(function(x) llfun(mu = 1, kp = x, lam = 0))
curve(kp_conditional_ll, 0, 20)
}
