% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vonMises.R
\name{logBesselI}
\alias{logBesselI}
\title{Log of the Bessel function}
\usage{
logBesselI(x, nu)
}
\arguments{
\item{x}{numeric, \eqn{\ge 0}.}

\item{nu}{numeric; The \emph{order} (maybe fractional!) of the
    corresponding Bessel function.}
}
\value{
The logarithm of the bessel function
}
\description{
This can be more numerically stable than taking the log directly.
}
