% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatscheletOptimization.R
\name{maxlikbat}
\alias{maxlikbat}
\title{Find maximum likelihood estimates for the inverse Batschelet distribution}
\usage{
maxlikbat(x, likfunbat_fun = likfuninvbat, weights, fixed_mu = NA,
  fixed_kp = NA, fixed_lam = NA, max_its = 20, kp_max = 100)
}
\arguments{
\item{x}{An set of angles in radians.}

\item{weights}{A vector of length \code{length(x)}, which gives importace
weights to be used for x.}

\item{fixed_mu}{If NA, the \code{mu} will be estimated as the mean direction.
Else, \code{mu} will be fixed at \code{fixed_mu}.}

\item{fixed_kp}{If NA, kp will be estimated. Else, a numeric giving the fixed
value of kp.}

\item{fixed_lam}{If NA, lam will be estimated. Else, a numeric giving the
fixed value of lam.}

\item{max_its}{Used only if neither \code{fixed_kp} nor \code{fixed_lam} is
provided. The maximum number of iterations for the Nelder-Mead optimization.}

\item{kp_max}{Used only if \code{fixed_lam} is provided. The maximum value
for kappa to search for.}
}
\value{
The maximum likelihood estimates for the \code{mu}, \code{kp}, and
  \code{lam}.
}
\description{
Find maximum likelihood estimates for the inverse Batschelet distribution
}
