% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InvBatscheletOptimization.R
\name{maxlikinvbat}
\alias{maxlikinvbat}
\title{Find maximum likelihood estimates for the inverse Batschelet distribution}
\usage{
maxlikinvbat(x, fixed_mu = NA)
}
\arguments{
\item{x}{An set of angles in radians.}

\item{fixed_mu}{If NA, the \code{mu} will be estimated as the mean direction. Else, \code{mu}
will be fixed at \code{fixed_mu}.}
}
\value{
The maximum likelihood estimates for the \code{mu}, \code{kp}, and \code{lam}.
}
\description{
Find maximum likelihood estimates for the inverse Batschelet distribution
}
