% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcBatschelet.R
\name{mcmcBatscheletMixture}
\alias{mcmcBatscheletMixture}
\title{MCMC sampling for Batschelet-type distributions.}
\usage{
mcmcBatscheletMixture(x, Q = 1000, burnin = 0, thin = 1, n_comp = 4,
  bat_type = "inverse", init_pmat = matrix(NA, n_comp, 4),
  fixed_pmat = matrix(NA, n_comp, 4), joint_kp_lam = FALSE, kp_bw = 1,
  lam_bw = 0.05, mu_logprior_fun = function(mu) -log(2 * pi),
  kp_logprior_fun = function(kp) 1, lam_logprior_fun = function(lam)
  -log(2), alph_prior_param = rep(1, n_comp), compute_variance = TRUE,
  compute_waic = FALSE, verbose = 0)
}
\arguments{
\item{x}{A numeric vector of angles, in radians}

\item{Q}{Integer; The number of iterations to return after taking burn in and
thinning into account.}

\item{burnin}{Integer; The number of (non-thinned) iterations to discard. No
burn in is performed by default.}

\item{thin}{Integer; Number of iterations to sample for each saved iteration.
Defaults to 1, which means no thinning.}

\item{n_comp}{Integer; Fixed number of components to estimate.}

\item{bat_type}{Either 'inverse' or 'power', the type of distribution to fit.
The two distributions are similar, but the power Batschelet distribution is
computationally much less demanding.}

\item{init_pmat}{A numeric matrix with \code{n_comp} rows and four columns,
corresponding to \eqn{\mu, \kappa, \lambda, \alpha}, in that order. Gives
starting values for the parameters. If any element is \code{NA}, it will be
given a default starting value. For \eqn{mu}, the default starting values
are equally spaced on the circle. For \eqn{\kappa}, the default starting
value is 5. For \eqn{\lambda}, the default starting value is 0, which
corresponds to the von Mises distribution. For \eqn{\alpha}, the default
starting value is \code{1/n_comp}.}

\item{fixed_pmat}{A numeric matrix with \code{n_comp} rows and four columns,
corresponding to \eqn{\mu, \kappa, \lambda, \alpha}, in that order. Any
element that is not \code{NA} in this matrix will be held constant at the
given value and not sampled.}

\item{joint_kp_lam}{Logical; If \code{TRUE}, the parameters \code{kp} and
\code{lam} are drawn jointly. This can be beneficial if these are strongly
correlated.}

\item{kp_bw}{Numeric; A tuning parameter for kappa proposals. If \code{kp_bw
== 1}, the chi-square distribution is used. Often, this distribution is too
wide, so this parameter can be set to \code{0 < kp_bw < 1} to use a gamma
proposal which has lower variance than the chi-square.}

\item{lam_bw}{Numeric; the maximum distance from the current lambda at which
uniform proposals are drawn.}

\item{mu_logprior_fun}{Function; A function with a single argument, which
returns the log of the prior probability of \eqn{\mu}. Defaults to a
uniform prior function.}

\item{kp_logprior_fun}{Function; A function with a single argument, which
returns the log of the prior probability of \eqn{\kappa}. Defaults to a
uniform prior function. In contrast to the other parameters, for
\eqn{\kappa} the constant (uniform) prior is improper.}

\item{lam_logprior_fun}{Function; A function with a single argument, which
returns the log of the prior probability of \eqn{\lambda}. Defaults to a
uniform prior function.}

\item{alph_prior_param}{Integer vector; The mixture weight parameter vector
\eqn{\alpha} is given its conjugate Dirichlet prior. The default is
\code{rep(1, n_comp)}, which is the noninformative uniform prior over the
\code{n_comp} simplex.}

\item{compute_variance}{Logical; Whether to add circular variance to the
returned mcmc sample.}

\item{compute_waic}{Logical; Whether to compute the WAIC. Can be
computationally demanding if \code{n * Q} is large.}

\item{verbose}{Integer up to 4; Determines the amount of printed debug
information.}
}
\value{
A numeric matrix of sampled parameter values.
}
\description{
MCMC sampling for Batschelet-type distributions.
}
\examples{
x <- rinvbatmix(100)
mcmcBatscheletMixture(x, Q = 10)
}
