% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatMixturesFitting.R
\name{multisummary.batmixmod}
\alias{multisummary.batmixmod}
\title{This function takes a list of bat_mix_mods, and provides a table that compares the fits.}
\usage{
multisummary.batmixmod(bm_mod_list, add_ci = TRUE, add_circ_sd = TRUE,
  add_circ_var = FALSE)
}
\arguments{
\item{bm_mod_list}{A list of \code{batmixmod} objects.}

\item{add_ci}{Logical; Whether to add confidence intervals.}

\item{add_circ_sd}{Logical; Whether to add circular sd.}

\item{add_circ_var}{Logical; Whether to add circular variance.}
}
\value{
A data frame with a summary.
}
\description{
This function takes a list of bat_mix_mods, and provides a table that compares the fits.
}
