% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plot.batmixmod}
\alias{plot.batmixmod}
\title{Plot a Batschelet-type mixture model}
\usage{
\method{plot}{batmixmod}(object, ...)
}
\arguments{
\item{...}{Additional arguments to be passed to \code{plot_batmixfit}.}

\item{bmm}{A \code{batmixmod} object.}
}
\value{
A \code{ggplot}.
}
\description{
Plot a Batschelet-type mixture model
}
\examples{
x <- rinvbatmix(50)
plot(fitbatmix(x, method = "EM"))

}
