% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plot_batmix_sample}
\alias{plot_batmix_sample}
\title{Plot a sample of Batschelet mixture parameter sets.}
\usage{
plot_batmix_sample(x, param, dbat_fun = dinvbat, plot_n = nrow(param),
  hist_transparancy = 0.3, dens_darkness = 20, bins = 100, res = 400,
  orderColor = FALSE)
}
\arguments{
\item{x}{An optional dataset of angles to be plotted as a histogram.}

\item{param}{A matrix of parameter sets.}

\item{dbat_fun}{A function; The pdf of the chosen Batschelet distribution.}

\item{plot_n}{Integer; the number of parameter rows to sample from the matrix
\code{param}. This is intended for MCMC for example, where we can take a
subsample of the parameter matrix to plot for speed.}

\item{hist_transparancy}{Numeric; The alpha value of the histogram of the data.}

\item{dens_darkness}{Numeric; Higher numbers result in less transparent densities plotted.}

\item{bins}{The number of bins to draw in the histogram.}

\item{res}{Number of points at which to evaluate the functions.}

\item{orderColor}{Logical; If \code{TRUE}, plotted pdfs get darker and redder
as they were sampled later.}
}
\value{
A ggplot.
}
\description{
Plot a sample of Batschelet mixture parameter sets.
}
\examples{


}
