% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plot_batmixfit}
\alias{plot_batmixfit}
\title{Plot a mixture of Batschelet type distributions.}
\usage{
plot_batmixfit(x, params, dbat_fun = dpowbat, bins = 100, res = 400,
  hist_alpha = 0.3)
}
\arguments{
\item{x}{Optional data vector of angles in radians to plot a histogram of.}

\item{params}{A matrix of parameters.}

\item{dbat_fun}{A function; the Batschelet function to use. Defaults to the Inverse Batschelet Function.}

\item{bins}{Integer; The number of bins to use in the optional histogram.}

\item{res}{Integer; The number of points at which to evaluate the density function.}
}
\value{
A ggplot.
}
\description{
Plot a mixture of Batschelet type distributions.
}
\examples{
plot_batmixfit(params = cbind(mu = c(-pi/2, 0, pi/2, pi), kp = 4, lam = c(-.9, .2, .8, 0), alph = .25))

}
