% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plot_movMF_as_batmix}
\alias{plot_movMF_as_batmix}
\title{plot_movMF_as_batmix}
\usage{
plot_movMF_as_batmix(m, ...)
}
\arguments{
\item{m}{A movMF results object.}

\item{...}{Additional arguments to be passed to \code{plot_batmixfit}.}
}
\value{
A ggplot.
}
\description{
Plot a result from movMF using the Batschelet mixture plotting functions.
}
\examples{

\dontrun{
  movMF_model <- movMF(cbind(runif(100, -1, 1), runif(100, -1, 1), 2), k = 4)
  plot_movMF_as_batmix(movMF_model)
}

}
