% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InvBatschelet.R
\name{rinvbat}
\alias{rinvbat}
\title{Random generation of inverse Batschelet distribution}
\usage{
rinvbat(n, mu = 0, kp = 1, lam = 0)
}
\description{
Random generation of inverse Batschelet distribution
}
