% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InvBatschelet.R
\name{s_lam}
\alias{s_lam}
\title{Inverse Batschelet subfunction}
\usage{
s_lam(x, lam)
}
\arguments{
\item{x}{An angle in radians.}

\item{lam}{The shape parameter.}
}
\value{
Numeric.
}
\description{
Inverse Batschelet subfunction
}
