% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcBatschelet.R
\name{vm_kp_jeffreys_prior}
\alias{vm_kp_jeffreys_prior}
\title{The Jeffreys prior for the von Mises distribution}
\usage{
vm_kp_jeffreys_prior(kp)
}
\arguments{
\item{kp}{Numeric;}
}
\value{
An unnormalized value.
}
\description{
The Jeffreys prior for the von Mises distribution
}
