% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InvBatschelet.R
\name{weight_fun_rinvbat}
\alias{weight_fun_rinvbat}
\title{Compute the weight function of the inverse Batschelet random generation function}
\usage{
weight_fun_rinvbat(x, lam)
}
\arguments{
\item{x}{An angle in radians.}

\item{lam}{The shape parameter (peakedness), -1 < \code{lam} < 1.}
}
\value{
Numeric.
}
\description{
Compute the weight function of the inverse Batschelet random generation function
}
