% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{AIC.batmixmod}
\alias{AIC.batmixmod}
\title{Extract AIC from Batmixmod}
\usage{
\method{AIC}{batmixmod}(object, ...)
}
\arguments{
\item{object}{A \code{batmixmod} object.}

\item{...}{Ignored.}
}
\value{
The numerical value of the AIC.
}
\description{
Extract AIC from Batmixmod
}
