% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{bridge_sampler.batmixmod}
\alias{bridge_sampler.batmixmod}
\title{Log Marginal Likelihood via Bridge Sampling}
\usage{
\method{bridge_sampler}{batmixmod}(samples, ...)
}
\arguments{
\item{samples}{A \code{batmixmod} object (from \code{\link{fitbatmix}}),
which contains the MCMC sample required for bridge sampling.}

\item{...}{Additional arguments to be passed to \code{bridge_sampler.matrix}.}
}
\description{
Obtain the log of the Marginal Likelihood through Bridge sampling. This is
useful to perform any Bayesian hypothesis test between any set of models.
}
\details{
More details are provided under
\code{\link[bridgesampling:bridge_sampler]{bridge_sampler}}.
}
\seealso{
\code{ \link[brms:bayes_factor]{bayes_factor},
  \link[brms:post_prob]{post_prob} }
}
