% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsOfBatParameters.R
\name{computeMeanResultantLengthBat}
\alias{computeMeanResultantLengthBat}
\title{The mean resultant length of a Batschelet-type distribution.}
\usage{
computeMeanResultantLengthBat(kp, lam, bat_type = "power")
}
\arguments{
\item{kp}{The concentration parameter.}

\item{lam}{The peakedness parameter.}

\item{bat_type}{Either 'inverse' or 'power', the type of distribution to
compute the resultant length for.}
}
\value{
The mean resultant length.
}
\description{
The mean resultant length of a Batschelet-type distribution.
}
