% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vonMises.R
\name{dvmkern}
\alias{dvmkern}
\alias{dvm}
\title{Kernel of the von Mises distribution.}
\usage{
dvmkern(x, mu = 0, kp = 1, log = FALSE)

dvm(x, mu = 0, kp = 1, log = FALSE)
}
\arguments{
\item{x}{An angle in radians for which to compute the pdf.}

\item{mu}{The mean direction.}

\item{kp}{The concentration parameter, \eqn{kappa}.}

\item{log}{Logical; If TRUE, return the log of the probability of x.}
}
\value{
The probability density of the von Mises distribution at x.
}
\description{
Kernel of the von Mises distribution.

Compute the probability density function of the von Mises distribution.
}
\section{Functions}{
\itemize{
\item \code{dvmkern}: The kernel of the von Mises distribution.
}}

\examples{
dvm(3)
dvm(3, -2.2, 10)
}
