% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invBatMixturesFitting.R
\name{fitbatmix}
\alias{fitbatmix}
\title{Fit mixtures of inverse or power Batschelet distributions.}
\usage{
fitbatmix(x, bat_type = "inverse", n_comp = 4, init_pmat = matrix(NA,
  n_comp, 4), fixed_pmat = matrix(NA, n_comp, 4), ll_tol = 1,
  max_its = 50, verbose = FALSE, optimization_its = 5)
}
\arguments{
\item{x}{A dataset of angles in radians.}

\item{n_comp}{The number of components to be used in the mixture. This is fixed, so it can not be
estimated.}

\item{init_pmat}{An \code{n_comp * 4} matrix, the initial values of the parameter matrix. The
parameters are ordered \code{mu}, \code{kp}, \code{lam} and then \code{alph}.}

\item{fixed_pmat}{An \code{n_comp * 4} matrix, containing a parameter matrix, with \code{NA} for
parameters to be estimated, and a numeric for each parameter that should be kept fixed to a
specific value.}

\item{ll_tol}{Numeric; the algorithm is stopped after the log-likelihood improves less than \code{ll_tol}.}

\item{max_its}{The maximum number of iterations.}

\item{verbose}{Logical; whether to print debug statements.}
}
\value{
A parameter matrix of results.
}
\description{
Fit mixtures of inverse or power Batschelet distributions.
}
\examples{
rinvbatmix

}
