% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatMixturesFitting.R
\name{fitbatmix}
\alias{fitbatmix}
\title{Fit a mixture of}
\usage{
fitbatmix(x, bat_type = "power", method = "bayes", n_comp = 4,
  init_pmat = matrix(NA, n_comp, 4), fixed_pmat = matrix(NA, n_comp, 4),
  verbose = FALSE, ...)
}
\arguments{
\item{x}{A dataset of angles in radians.}

\item{bat_type}{Character; Either "power" or "inverse", denoting the type of
Batschelet distribution to employ.}

\item{method}{Character; One of "bayes", "EM", or "boot". The method of obtaining a fit.}

\item{n_comp}{The number of components to be used in the mixture. This is
fixed, so it can not be estimated.}

\item{init_pmat}{An \code{n_comp * 4} matrix, the initial values of the
parameter matrix. The parameters are ordered \code{mu}, \code{kp},
\code{lam} and then \code{alph}.}

\item{fixed_pmat}{An \code{n_comp * 4} matrix, containing a parameter matrix,
with \code{NA} for parameters to be estimated, and a numeric for each
parameter that should be kept fixed to a specific value.}

\item{verbose}{Logical; whether to print debug statements.'}

\item{...}{Additional arguments to be passed to the selected \code{method}.}
}
\value{
An object of class 'batmixmod'.
}
\description{
This is the main function of the package \code{flexcircmix}, and functions as
an interface to fit mixtures of Batschelet-type distributions, using
frequentist or Bayesian methods.
}
\examples{


}
