% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InvBatschelet.R
\name{likinvbat}
\alias{likinvbat}
\title{Likelihood function of inverse Batschelet.}
\usage{
likinvbat(x, mu, kp, lam, weights = rep(1, length(x)), log = TRUE)
}
\description{
@describeIn likfuninvbat
 @export
}
