% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatscheletOptimization.R
\name{maxlikbat}
\alias{maxlikbat}
\title{Find maximum likelihood estimates for the inverse Batschelet distribution}
\usage{
maxlikbat(x, likfunbat_fun = likfuninvbat, weights, fixed_mu = NA,
  fixed_kp = NA, fixed_lam = NA, init_kp = 1, init_lam = 0,
  max_its = 20, kp_max = 1000)
}
\arguments{
\item{x}{An set of angles in radians.}

\item{likfunbat_fun}{A function that returns a function, which is the desired
likelihood function to maximize over. Either \code{likfuninvbat} (the
default) or \code{likfunpowbat}.}

\item{weights}{A vector of length \code{length(x)}, which gives importance
weights to be used for x.}

\item{fixed_mu}{If NA, the \code{mu} will be estimated as the mean direction.
Else, \code{mu} will be fixed at \code{fixed_mu}.}

\item{fixed_kp}{If NA, kp will be estimated. Else, a numeric giving the fixed
value of kp.}

\item{fixed_lam}{If NA, lam will be estimated. Else, a numeric giving the
fixed value of lam.}

\item{init_kp}{An initial value for \eqn{\kappa}, only used if neither
\eqn{\kappa} or \eqn{\lambda} is fixed.}

\item{init_lam}{An initial value for \eqn{\lambda}, only used if neither
\eqn{\kappa} or \eqn{\lambda} is fixed.}

\item{max_its}{Used only if neither \code{fixed_kp} nor \code{fixed_lam} is
provided. The maximum number of iterations for the Nelder-Mead
optimization.}

\item{kp_max}{Used only if \code{fixed_lam} is provided. The maximum value
for kappa to search for.}
}
\value{
The maximum likelihood estimates for the \code{mu}, \code{kp}, and
  \code{lam}.
}
\description{
Find maximum likelihood estimates for the inverse Batschelet distribution
}
