% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plot_batmix_sample}
\alias{plot_batmix_sample}
\title{Plot a sample of Batschelet mixture parameter sets.}
\usage{
plot_batmix_sample(x = NA, param, dbat_fun = dinvbat, bins = 100,
  res = 400, orderColor = TRUE)
}
\arguments{
\item{x}{An optional dataset of angles to be plotted as a histogram.}

\item{param}{A matrix of parameter sets.}

\item{dbat_fun}{A function; The pdf of the chosen Batschelet distribution.}

\item{bins}{The number of bins to draw in the histogram.}

\item{res}{Number of points at which to evaluate the functions.}

\item{orderColor}{Logical; If \code{TRUE}, plotted pdfs get darker and redder as they were sampled later.}
}
\value{
A ggplot.
}
\description{
Plot a sample of Batschelet mixture parameter sets.
}
\examples{


}
