% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatMixturesFitting.R
\name{print.batmixmod}
\alias{print.batmixmod}
\title{Printing function for \code{batmixmod} objects.}
\usage{
\method{print}{batmixmod}(x, ...)
}
\arguments{
\item{...}{Further arguments to be passed to print methods.}

\item{bm_mod}{A \code{batmixmod} object.}
}
\description{
Printing function for \code{batmixmod} objects.
}
