% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InvBatschelet.R
\name{s_lam}
\alias{s_lam}
\alias{s_lam_inv}
\alias{s_lam_inv_vec}
\title{Inverse Batschelet subfunction}
\usage{
s_lam(x, lam)

s_lam_inv(x, lam)

s_lam_inv_vec(x, lam)
}
\arguments{
\item{x}{An angle in radians.}

\item{lam}{The shape parameter.}
}
\value{
Numeric.
}
\description{
Inverse Batschelet subfunction

Inverse Batschelet inverse subfunction

Vectorized version of the inversion of s_lam
}
\section{Functions}{
\itemize{
\item \code{s_lam_inv}: 

\item \code{s_lam_inv_vec}: 
}}

