% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InvBatschelet.R
\name{t_lam}
\alias{t_lam}
\alias{t_lam_inv}
\title{Inverse Batschelet function}
\usage{
t_lam(x, lam)

t_lam_inv(x, lam)
}
\arguments{
\item{x}{An angle in radians.}

\item{lam}{The shape parameter.}
}
\value{
The transformed angle x.
}
\description{
Function used to generate different shapes for the inverse Batschelet distribution.

Inverse Batschelet inverse function
}
\section{Functions}{
\itemize{
\item \code{t_lam_inv}: 
}}

\examples{

# Reproduce Figure 5 in Jones & Pewsey
curve(t_lam(x, 0), -pi, pi, asp = 1)
for (i in seq(-1, 1, by = 1/3)) {
  curve(t_lam(x, i), -pi, pi, add = TRUE, lty = "dotted")
}
}
